requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
greenlet==3.2.3
asyncpg==0.30.0